Got primitive height syncing working.
The PreferenceKey only seems to work if every view has the exact same signature, meaning you can’t have an image background for one cell and nothing for another. Even an if bool { SubView() } construction breaks the PreferenceKey. This CL changes all of the ColorSwatch signatures to be identical (stuff like setting the supported OS tag to opacity 0 when drawing an empty cell over omitting the tag altogether).

It works great on iOS/iPadOS & tvOS. It doesn’t work on macOS but I want to comit this before figuring out why. It also isn’t changing the *background* of the cells, just the frame.

Also the HeightSyncedRow code is ripped straight out of of a Swift by Sundell article. I intend to refactor it into FlowableContentGrid before finishing.

- Also removed the layoutPriority from SupportedOSTagView, as it is no longer needed with superior column width calculations in place.